Now tagging only appears when enabled from admin.

+++ controllers/admin/blog.php
@@ -66,8 +66,11 @@ class Blog extends MY_Controller
 	public function action_create()
 	{
 		$data = $this->input->post('starter_article');
-		$tags = $data['tags'];
-		unset($data['tags']);
+		if ($this->module->setting('include_tags'))
+		{
+			$tags = $data['tags'];
+			unset($data['tags']);
+		}
 		$data['user_id'] = $this->current_user->id;
 		$article = $this->article_model->create($data);
 				
@@ -78,15 +81,18 @@ class Blog extends MY_Controller
 		}
 		else
 		{
-			//add tags
-			$set = array();
-			$tags = explode(',', $tags);
-			foreach ($tags as $tag)
+			if ($this->module->setting('include_tags'))
 			{
-				$this->article_tag_model->create(array(
-					'starter_article_id' => $article->id,
-					'name' => $tag
-				));
+				//add tags
+				$set = array();
+				$tags = explode(',', $tags);
+				foreach ($tags as $tag)
+				{
+					$this->article_tag_model->create(array(
+						'starter_article_id' => $article->id,
+						'name' => $tag
+					));
+				}
 			}
 			
 			flash('notice', 'Article was created successfully.');
@@ -106,8 +112,11 @@ class Blog extends MY_Controller
 	public function action_update($id)
 	{
 		$data = $this->input->post('starter_article');
-		$tags = $data['tags'];
-		unset($data['tags']);
+		if ($this->module->setting('include_tags'))
+		{
+			$tags = $data['tags'];
+			unset($data['tags']);
+		}
 		$data['user_id'] = $this->current_user->id;
 		$article = $this->article_model->update($id, $data);
 				
@@ -119,15 +128,21 @@ class Blog extends MY_Controller
 		else
 		{
 			//add tags
-			foreach ($article->tags->all() as $tag) $tag->destroy();
-			$set = array();
-			$tags = explode(',', $tags);
-			foreach ($tags as $tag)
+			if ($this->module->setting('include_tags'))
 			{
-				$this->article_tag_model->create(array(
-					'starter_article_id' => $article->id,
-					'name' => $tag
-				));
+				//remove tags
+				foreach ($article->tags->all() as $tag) $tag->destroy();
+				
+				//add tags
+				$set = array();
+				$tags = explode(',', $tags);
+				foreach ($tags as $tag)
+				{
+					$this->article_tag_model->create(array(
+						'starter_article_id' => $article->id,
+						'name' => $tag
+					));
+				}
 			}
 			flash('notice', 'Article was updated successfully.');
 		}
